workspaces/rabbit-roboclaw/pyproject.toml
[project]
name = "rabbit-roboclaw"
version = "0.0.1"
requires-python = ">=3.12,<3.13"
dependencies = ["nats-py", "pyserial"]
